PANIC moment: checked exceptions going through a @Transactional proxy do NOT kill the current Tx. WHO THE HACK THOUGHT THIS IS A GOOD IDEA?
it was 2005. EJB 2.1 (JavaEE) was reigning the Java planet. 80% of Java dev working for large companie were using one of the WORST STANDARD EVER ON PLANET EARTH.

A group of nomads create Spring to fight the shit in the standard.

They wanted to ease the transition from EJB to Spring. They wanted to "steal" developers.
Since then up to now, spring was always ahead of JavaEE 2-4 years. because JavaEE is a standard, backed by a consortium of companies. they had to agree -> slower.

So they replicated the mistakes of @TransactionAttribute (EJB)