pitfalls when method interception (ie. annotations on methods) don't work anymore
!!! WORST :
It seems like you are using the @Cacheable("sums") annotation on your sum method, which is a feature from the Spring framework to cache the result of the method call. However, please note this caching will work only when the method is called as a Spring bean method (means the method is called from another Spring bean not from the same class).